requirements.txt
asgiref==3.8.1
sqlparse==0.5.3
psycopg2>=2.9.10,<2.10
cryptography>=44.0.2,<44.1
Django>=5.2,<5.3
djangorestframework>=3.16.0,<3.17.0
drf-spectacular>=0.28.0,<0.29
djangorestframework-simplejwt>=5.5.0,<5.6

